+++ R/ee_download.R
@@ -44,8 +44,9 @@
 #' deprecated such as 'crs_transform', 'driveFolder', and 'driveFileNamePrefix'.
 #'
 #' @return An unstarted Task that exports the image to Drive.
+#' @family image export task creator
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' library(rgee)
 #' library(stars)
 #' library(sf)
@@ -196,8 +197,9 @@ ee_image_to_drive <- function(image,
 #' deprecated such as 'crs_transform'.
 #'
 #' @return An unstarted Task that exports the image to Google Cloud Storage.
+#' @family image export task creator
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' library(rgee)
 #' library(stars)
 #' library(sf)
@@ -306,9 +308,9 @@ ee_image_to_gcs <- function(image,
   )
 }
 
-#' Creates a task to export an EE Image to an EE Asset.
+#' Creates a task to export an EE Image to their EE Assets.
 #'
-#' Creates a task to export an EE Image to an EE Asset.
+#' Creates a task to export an EE Image to their EE Assets.
 #' This function is a wrapper around \code{ee$batch$Export$image$toAsset(...)}.
 #'
 #'
@@ -342,9 +344,10 @@ ee_image_to_gcs <- function(image,
 #' other keyword arguments that may have been deprecated such
 #' as 'crs_transform'.
 #'
-#' @return An unstarted Task that exports the image to Earth Engine Asset
+#' @return An unstarted task
+#' @family image export task creator
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' library(rgee)
 #' library(stars)
 #' library(sf)
@@ -461,8 +464,9 @@ ee_image_to_asset <- function(image,
 #' 'driveFileNamePrefix'.
 #'
 #' @return An unstarted Task that exports the table to Google Drive.
+#' @family vector export task creator
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' library(rgee)
 #' library(stars)
 #' library(sf)
@@ -546,8 +550,9 @@ ee_table_to_drive <- function(collection,
 #' that may have been deprecated such as 'outputBucket'.
 #'
 #' @return An unstarted Task that exports the table to Google Cloud Storage.
+#' @family vector export task creator
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' library(rgee)
 #' library(stars)
 #' library(sf)
@@ -629,8 +634,9 @@ ee_table_to_gcs <- function(collection,
 #' it exists.
 #'
 #' @return An unstarted Task that exports the table to Earth Engine Asset.
+#' @family vector export task creator
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' library(rgee)
 #' library(stars)
 #' library(sf)
@@ -713,13 +719,14 @@ ee_table_to_asset <- function(collection,
 #' will download just the last file saved in Google Drive while with "all" all
 #' files will be downloaded.
 #'
-#' @return filename character vector.
-#'
 #' @importFrom stars read_stars
 #' @importFrom utils menu
 #'
+#' @return filename character vector.
+#' @family generic download functions
+#'
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' library(rgee)
 #' library(stars)
 #' library(sf)
@@ -924,9 +931,10 @@ ee_drive_to_local <- function(task,
 #' \code{ee$batch$Export$*$toCloudStorage(...)$start()$status()}.
 #'
 #' @return filename character vector.
+#' @family generic download functions
 #'
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' library(rgee)
 #' library(stars)
 #' library(sf)