+++ R/ee_help.R
@@ -2,8 +2,9 @@
 #' @param eeobject Earth Engine Object to print documentation.
 #' @param browser Logical. Display documentation in the browser.
 #' @importFrom reticulate py_function_docs
+#' @family helper functions
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' library(rgee)
 #'
 #' ee_reattach()