+++ R/ee_image.R
@@ -37,9 +37,10 @@
 #' a look at the \href{https://developers.google.com/earth-engine/exporting}{Google
 #' Earth Engine Guide - Export data}.
 #' @return A stars-proxy object
+#' @family image download functions
 #' @export
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' library(rgee)
 #'
 #' ee_reattach()
@@ -159,8 +160,9 @@ ee_as_stars <- function(image,
 #' a look at the \href{https://developers.google.com/earth-engine/exporting}{Google
 #' Earth Engine Guide - Export data}.
 #' @return A RasterStack object
+#' @family image download functions
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' library(rgee)
 #'
 #' ee_reattach()
@@ -265,9 +267,9 @@ ee_as_raster  <- function(image,
 #' @importFrom geojsonio geojson_json
 #'
 #' @return An ee$Image object
-#'
+#' @family image upload functions
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' library(rgee)
 #' library(stars)
 #' ee_Initialize(gcs = TRUE)
@@ -354,9 +356,10 @@ stars_as_ee <- function(x,
 #' @importFrom geojsonio geojson_json
 #'
 #' @return An ee$Image object
+#' @family image upload functions
 #'
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' library(rgee)
 #' library(stars)
 #' ee_Initialize(gcs = TRUE)
@@ -799,9 +802,11 @@ ee_image_local <- function(image,
 #' (ignored if \code{getsize} is FALSE). By default is 20
 #' @param quiet Logical. Suppress info message
 #' @importFrom sf st_transform
-#' @return A list of parameters
+#' @return A list containing information about the number of rows (nrow),
+#' number of columns (ncol), total number of pixels (total_pixel), and image
+#' size (image_size).
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' library(rgee)
 #'
 #' ee_reattach()